Player Struct, Abilities
A player's data is now stored in a struct, and the game state is made up of an array of 8 players. There's always 8 players in the array (since it has less variable memory usage), but only a certain number of players are online at a time.

Each player has a color, coordinates, an ability, and speed, that can be different from everyone else.

All of the game logic is now stored in the game.rs files, and the main.rs file will be made up of stuff like rendering the game's state.

The game also runs on 60 fps now, and uses VSync