Added shooting to the game
So bullets are these things called projectiles, with attributes similar to a player's such as x and y coords, speed, and direction it's moving in. Each player will have a different weapon, which will create projectiles of different speeds and stuff.

Also added a quick demo for bots. Basically each bot is given an index in the array, and is also given information on all the player data every frame, and then chooses to move in a certain direction because of it. Eventually, the bots will have control over using their abilities, shooting, and other fun stuff, but for now I'm keeping it basic. The current bot called bounce just moves back and forth across the screen, as a demo of how bots work. All bot code is stored in bots.rs, and players will each be given a function that eventually has to give an output of what direction it wants to move.