Bullet gen alg, reload key, beginning of network
First off, bullets now generate in a different way. If depending on the
direction the player is moving in, the bullet will spawn in a different
place so that it is slightly in front of the player.

Added a reload key, so that players can now shoot and move easier

Also made an empty net.rs, where the network implementation will be.
Will probably use a Rust library called Naia, but who knows yet