Added zoom + optimized drawing
I've been worried about getting the players to look the exact same way at different screen resolutions, and adding a "zoom" to where you can see farther or closer is going to be part of that.

Also, the game now only draws sprites that are within the camera's viewport, which greatly optimizes the program.