Added speedball weapon + GGEZ 0.6-rc1
Firstly, added the speedball weapon, both as a way of having more unique guns to expand the Gun struct (with width and height), as well as because of the fact that I thought it'd be fun. The projectile's speed and size increase over the distance it travels.

Also upgraded GGEZ to version 0.6, for two reasons. One, we can now run the game on Rust stable, thank god. It was annoying as hell having to use a slightly old version of Rust. Secondly, it allows our codebase to be ready for the official stable release of 0.6.0, which should be faster and use later versions of dependencies.