Added simple map objects + fixed viewport
Firstly, I added the ability to add simple map objects. They can only be white  at the moment, they aren't dynamically loading in, and there's no way to export or import them, but still, they work.

I also fixed a bug with the new and improved dynamic viewport where if the top left corner of a rectangle wasn't in view, it wouldn't show the object.