Fixed mouse aiming bug
So basically, once the player was past the 400 x or 300 y mark, the gun
would always shoot to the top left corner because of the trig math I
did, so I just slightly modified how the gun aims (great explanation I
know). I'm tired, I spent 3 hours debugging this because I'm an idiot.