Projectile Rectangle Caching w. Sprite Batches
Drawing a new sprite requires creating a Mesh, creating a sprite, and then discarding both of those, which is a relatively expensive process. As a result, creating hundreds of sprites (for example, bullets) causes the game to lag, even on a decent laptop like mine. To solve this, I made it so the game will now use sprite batches for every projectile size, so that the game will do all the expensive stuff one time and then reuse the same sprite cheaply over and over, so now there can be literally be thousands of bullets on screen with very little downsides. This could be further optimized by having the sprite batch be a mutable static reference, but a. that's complicated lol and b. it would require unsafe, so we'd have to do some stuff with mutexes which is also complicated, so for now this is a good enough optimization. Doing the same with players will also give a minor boost to performance.