Added some HUD info + ability charges
Firstly, the screen now shows the health of each player on the screen (take notes ladies and gentleman, some wild stuff here.) I'm planning on adding some more information but I'm more focused on bug fixes and other things.

Secondly, abilities now have a thing called a 'charge', allowing a player to use an aiblity immediately when it's ready if they need it at that moment, or waiting a little longer and using a more powerful / longer lasting ability. For example, on the teleportation guy, waiting 2.5 seconds (300 ticks) causes the player to teleport a distance of 150 pixels, but waiting a full 5 seconds will double that (it's supposed to let you teleport twice, not double your distance, but I'll fix that later lol. It's because of the fact that the player can be pressing the ability button for multiple ticks on accident).

Also did some minor QOL stuff, like making it so players have to be brighter colors (it's hard to see them with the black background), and some clippy fixes for more efficient code.