More general rectangle spritebatch, maps now use spritebatch
So I decided that using a spritebatch for every rectangle would be much more useful than having a spritebatch for projectiles and maps, since that way there can be alot more caching. In the future, we should probably split stuff like large map objects into 10 by 10 grids, so that we can reuse a lot of spritebatches instead of making one for an object of each size.