Added the burst rifle
The burst rifle fires 3 shots in quick succession, with a 50 milisecond gap in between each shot. Like most guns that have been added so far, I added it just to expand how the gun engine works, and to make future guns that aren't basically just a pistol easier to add.