Added the assault rifle
Why? Because I'm lazy and I feel like having an automatic rifle is good
for the game lol. It's pretty strong right now with a very high dps if
you can aim, but slow travel time. Should probably balance it at some
point but it's nice to just have it out there.