Added the wall ability
Creative name I know. This abiliity allows the player to put up a small green wall as cover (or for escaping a chasing enemy, blocking off an area, etc). The cooldown is currently very very low for how strong this ability is imho, and I think that we need to add destructible walls at some point so this ability isn't incredibly over powered.