Added the engineer ability
The engineer can reload in half the time, as well as having their
projectiles move at 1.25x the speed of a normal projectile. Both of
these abilities are passive, meaning if they press the 'e' button,
nothing will happen (though the game hopefully won't bug out lol). For
the speedball weapon, the projectile's size and speed is a little faster
and bigger. I feel like making this the default character would be good,
since it's similar enough to other shooters that hopefully players will
be able to dip their feet into the game.