Moved all game logic into its own crate
This took like 2 hours lol. So to help with #1, #8, and #9, I tried to move all of the game's logic into its own crate, located at /game_logic at the Git repo root. By doing this, all of the games logic can be compiled into WASM, which is useful for porting the game to browsers, or into a .so file, useful for porting the game to iOS or Android. The only part of the code that would have to be rewritten is the main.rs file, (aka the game crate). The game_logic crate can be reused across any platform now!