Spawn points + num of players + minor improvements to tiles
it's now possible to set spawn points all around the map, as well as setting the number of players the map will have depending on the number of spawn points. Also made it so that you can have walls that aren't collidable (currently only used for player spawns, could be used for floors.)

Also made some minor improvemnets to the default tileset so you can actually pick red green and blue.