Added flooring, improved mapping color
Firstly, the ability to add floors was massively improved, and now the map has a nice dark cyan color. 

Secondly, there was an issue with serde_json where it would interpret the array given to it into a tuple of floating points, which is not what I wanted at all lol. So now instead, I moved all the floating point work into the custom.js file, where it will just turn the RGBA colors into RGBA but in a float between 0 and 1, where it is interpreted correctly.