Fixed map rendering + camera
So after messing with the map, the js code, and serde to try to figure out why the game's map would render weirdly, I figured out that because of some internal algorithm serde uses, the map's objects wouldn't always come in order, so sometimes the floor would be rendered over some objects. To try and counter this, I'm currently just putting the floor on the objects layer, and putting the player spawn points in its own separate layer (since it has an opacity of 0, it doesnt matter where it gets spawned.

Also made a minor change to how the camera moved so that it's impossible to look off the map.