Maps now save multiple layers
So the map was rendering weirdly and I was fairly sure it was an issue with the .js script, and it turns out that I was right. I wasn't allocating enough space in the u8 array for multiple layers, so I'm instead just having a regular array that I turn into a byte array.