Switched from rand to getrandom
So apparently, rand doesn't support WASM, which is a big problem for a game that's going to be playable on any platform. After 2 minutes of Googling I found out that the getrandom crate can support almost every platform under the sun (including Mac and iOS), so I switched to it.

It's slightly more annoying in that it gives you random bytes, though it also gives you more flexibility.

I also think I finally filled in all the floor holes in the map.