Added checksum to map files + disabled debug info
So I got around to adding a checksum to the mapfiles, to self test for corrupted maps, and it seems to be working fine. It doesn't prevent users their own maps or anything, which is intended. I also used crc32 instead of a cryptographic checksum like sha256 just because it's really fast, like amd64 chips include it as an instruction. It also builds to every (almost) every platform since it includes no_std support, which is nice. If it turns out that for some reason it won't build to WASM, it'd only take a few hours to rewrite it since it's pretty simple math.

Also disabled all debug information when compiling, since neither me nor luke ever use a debugger in our builds (I hope lol). This has honestly greatly improved compile times, while not changing any feedback the compiler gives or panics give.