Added a settings screen
Now there are three screens in the game, and in the settings screen you can set the keybindings (though it doesn't save after restarting the game at the moment).

Switched the MainState view value to using an enum, and also added an entire View struct, and will probably do the same for the game.