Added a game "log", game_state struct
So in the bottom right corner of the game, there's now a thing called a "log", that just has the events of the game going on. To test it, I made it so whenever a player took damage, it was added to the long. In the actual game, however, it'll probably for big events like the death of a player, a flag being captured, etc.

Also did a little bit for #15 by removing the Copy and Clone traits for Player.