Fixed multiple bullets bug + MainState time
There was a bug in lib.rs since I accidentally redefined i twice, where if there were more bullets on screen than players, and a bullet hit a player, then the game would crash.

Also made it so that in main.rs, there was a single universal time, instead of multiple calls to `current_time()`, both for efficiency and for consistency.