Map Drawing
Ported over the relevant parts of map.rs for drawing the map to Bevy. Right now there's absolutely 0 sprite batching involved, so it runs absolutely horribly, but I'm planning on adding it in the next commit. The map also starts in the corner of the screen, though it does look good and like the old map did.