Decreased map size + increased game frame rate/efficiency
So one of the first things that I noticed when porting over our old map was just how incredibly slow it ran. Looking into it, I found out that Bevy doesn't currently support sprite batching >:(. 

Firstly, I removed the current floor and then added the background part of the map. This helped somewhat (the game was at a playable framerate), though it still stuttered. I also increased the optimization level by 1, which also helped, but it still of course isn't enough for me.

I then finally got to work on the object width thing in #12, and got it working, which significantly cut down on the number of sprites and in turn significantly improved the game's frame rate (it's now back to 60 fps). This also had the nice side effect of decreasing the map's overall size down significantly. Further improvements could be made by actually working on the height attribute of objects as well, making it so that the game would render a few dozen objects instead of hundreds, but I'm happy enough for now.

Also switched from Bevy 0.5 to the development version of Bevy, since they include some interesting speed performances. I have it locked to a certain Git commit in case of breaking changes, though I'll try to update it as often as possible.