Fixed map object collision
So Bevy internally uses a different size for sprites then what our game uses, so I basically just made it so that when calculating collisions, it uses Bevy's internal sprite size.

Also switched on the optimizations, though it draws maps weirdly, so it'll be optimized further in a later commit.

Also now using Vectors for storing the coordinates and height of map objects and players, since it reduces boilerplate code.