Map object storage + reduction in map + drawing objects fix
The Tiled Object is used instead of a ton of individual tiles, which both significantly reduces the number of sprites (resulting in a more efficient game and much smaller file size), as well as reducing the complexity of the code by not having to do any RLE stuff. The map size is down to 200 bytes!!! This will also let us use customized floor colors in the future, if need be, since there will be much much less sprites to draw. Combined with sprite-culling, the game should run at a smooth 60 fps on all platforms! 

None of these optimizations would have been done if it wasn't for stupid Bevy not having any stupid sprite batching, so that's honestly great!

Also modified how Bevy objects were draw to account for the center of the screen being at (0, 0), so no more weird drawing issues / hitbox issues!