Fixed collision function
So the way that Bevy has it's collision function is really weird, and basically it only works some of the time. I just copied and pasted the relevant parts of the code that I wanted (the ones that adjust the translation of the object's coordinates, you can see a similar thing in the from_bin function of the Map struct).