Shooting cooldown times + disappearing bullets + pistol
So I added cool down times to bullets, using Bevy's absolutely fantastic timer system. It can be used in the future for ability cooldowns as well as reloading, and anything else that needs to run at a specific rate.
Bullets also disappear when running into walls or when they move past 900 pixels.
Also added the pistol to the game, or at least parts of it, I'm working on getting all the features of guns in, which shouldn't take long.
Since the `main.rs` file is getting pretty big, I made a `player_input.rs` file to help with readability.
Thanks to the patch feature of Cargo.toml, we're now using the latest git commit (at the time of writing) of Bevy, in order to benefit from all the performance improvements and new features.