Reloading + UI + Main branch Bevy + Updated LZ4 + Refactors
So to start with, I finally took the time to add reloading in the game (once again using Bevy timers). I have to use a few separate systems, one to start reloading, one to tick the timer, and one to know when reloading is finished, to get a full reload working, but thankfully I can just reuse the last two for all players reloading (by providing a player ID) as well as reusing it for abilities. #17
There's also now a real UI for seeing how much ammo is in your gun, whether you're reloading, etc. Because of the `frustrum_culling` bug where it doesn't work with multiple cameras (the UI camera is separate), we unfortunately have to disable it. We should probably add our own implementation of sprite culling, though currently with hundreds of bullets on screen the game runs perfectly, I think it's only when running hundreds of thousands of individual sprites do we see performance drops. Nonetheless, we should still work on it.
We're also now using the main branch of Bevy for all of the latest performance improvements and stuff using cargo patch. I'd been holding off on it since unfortuanetly a bug prevented me from it, but some of the kind people over on the Bevy discord got a commit going to fix it in a half hour.
I normally don't comment on LZ4 updates, since normally its just updating the git repo, but since it's a big enough change to where I had to mess with the decompression code, I figured it was wroth it. Frame decompression is now gated by a feature, and it includes its own decompression function, which we're using instead.
Minor refactors of some files, comments, and general improvements to comprehensibility of code were made.