Changed color of engineer bullets + stim bug fix
So I made it so that the engineer's bullets would be pink, so as a sort of nerf ot signify that the player firing at them can be much more aggressive (considering halved reload time) and also since it looks cool.

Fixed a minor bug where engineer or stim could use the stim's abilities.