Basic main menu
So I finally got a basic main menu for the game up and running. You can only play the game, with no option to do stuff like change key bindings, but hey it's a start.

Unfortunately, because of how States work, I have to remove the code for setting a FixedTimestep, which is bad, especially for WASM. Setting a sepcified frame rate is important for the game not being very slow or running like garbage. There's a way to combine States and with_run_critera, but I don't feel like doing it right now, so I'm gonna make a Github issue.