Settings menu
So I finally got around to adding a settings menu, as well as figuring out how the flexbox format actually works, so now it'll actually center properly on all screen sizes.

This completes #17 settings and start screen, which was honestly the most difficult part. The rest should be a cakewalk.