Projectile-player collisions + damage
So projectiles can now collide with players, doing damage depending on the gun. Health is currently meaningless, but later I'm going to make it so that players with 0 health die for a few seconds, and then respawn.

Also fiex a tiny bug where if you exited the settings menu and then re=entered it, the game would crash. This is because I wasn't properly cleaning up the menu, so it would try to insert the same entity twice (resulting in a crash).