WebRTC Client-Server model
As noted in #4, the peer to peer model is being annoying, so I'm putting it on the back seat for a bit (until I figure out a better way to get it to work of course). Got a better ping pong example set up.

I had to add a couple of patches, since unfortuanetly, naia is currently having issues running in Firefox as is. It should work perfectly now, however.

To try it out, just run `cargo make serve` and visit that web page, and also run `cargo make run` at the same time in a different terminal to get the game build up and running, every second or so, you should see ping pong in the terminal log.