Proper hosting and joining UI
So due to an issue `naia-client-socket`, only native builds can host, and only web builds can join (for now). It's because naia doesn't have a proper WebRTC client for non-wasm builds :'(. We're gonna have to have a whole separate port for UDP messages, but we won't have to add any more network logic.

Also used a lot of conditional compilation, since a lot of the code would just be bloat if compiled for web or native, or vice versa.