Network shooting
When a player shoots, after the checks to see if they can shoot and all that, it sends the message out to all other player that are connected. Since the game is currently just trusting the client on whether they can or can't shoot, no verification is (currently) done.

Unfortunately, I have to use a very specific version of Bevy at the moment, since they're using an upgraded version of WGPU