Naia 0.6 + Multiplayer Aim Bug Fix
To get rid of the stupid patch that was in the deps directory, I made a pull request of the pull request that I was already using in order to use the latest version of Naia (0.6 for client, 0.5 for server), which not only includes a bug fix for the Firefox issue I had a long with a few other nice things.

Also fixed a minor bug fix for non player 1 characters where there bullets would go in the wrong direction, due to the issue with bullets needing to travel backwards to move left.