Added support for 3+ players + larger players
So how the game is currently working is that when a player tries to join, they request an ID from the central server, which when they receive is used as their regular ID. When a player moves or shoots, they include their ID, and then whenever the server receives an event, they broadcast it out.

Also made players larger in general, so that they're easier to hit and so they don't feel like they move so quickly.

Did an upgrade to Cargo.lock to support the new versions of a lot of different dependencies, including LZ4 Flex which means I had to change how it worked slightly.

In theory, the game could support a maximum of 256 players, since that's the maximum possible amount of unique ID's. A map would have to support 256 players, of course, and I'm not sure how the server could handle it jaja.