Fixed shooting bug + made logging an event
So shooting to the right wasn't possible if you were player 2, because of some weird stuff with calculating the negative speed twice, so I made it so that whether the bullet would be shot to the left was calculated locally, and then sent to the server.

Logging is also now an event, so that there's way less code involved for each log.

Finally, just for @Susorodni, I changed the default ability button to Q.