Added automatic Sccache usage
So right now, even the tiniest update to the code causes pretty long re-compilations. Sccache helps to mitigate this, by caching compilation results, so if the compiler asks to compile a certain library again, it'll just give back the cached compiled library, saving time.

I didn't know that cargo-make could do scripting, so that was cool to find out. The program just checks if sccache already exists on your system, and if it doesn't, it installs it.