Added SIMD support to WASM + Bulk Memory Ops
While trying to squeeze as much performance out of the compiler flags as possible, I added SIMD support and bulk memory operation support to the WASM binary, and found that it generally seemed to improve speed (though I'm unsure as to whether bulk memory operations or SIMD did more). 

Unfortunately, not all browsers support SIMD or even bulk memory ops so I used a library from Google and a simple if statement to detect support for a few browser features. Since Firefox and Brave seem to support bulk memory operations already, I just forced it to be necessary, and the binary won't load without it. Since SIMD is behind a feature flag, however, I made it so that two binaries would be compiled, one with SIMD support and one without it. Honestly I partially added SIMD support since I have it enabled, and I want debug builds to run faster.

#18