Added the hacker ability
The hacker ability picks a random player that isn't the hacker, and forces them to use their ability,, as well as forcing their ammo to be 0. I also made it so that every online player is kept using the OnlinePlayerID resource, that's kept as a binary tree.

I decided to specifically use a binary tree because of it's speed and the fact that it automatically sorts items, and sorted items are faster than a vector.

Since lz4_flex finally got merged, I'm using the latest version of the main branch.

Also, the issue with the latest version of nightly got fixed in packed_simd_2, so we're finally switching over to that.

I added a serve-fast-simd option since I only do SIMD builds lol.