General QOL Improvements
Firstly, I changed the protocol for sending player location data to send the health of the player, as well as the direction the player is facing, so that deaths sync up better.

Added the death event, so that when a player dies, it's easier to do stuff like change their visibility and log stuff, as opposed to having logs in several places.

I also made it so that players would get a random gun on startup.

Added deny clippy all to all files so that my IDE would give me more info.

Did some general changes to the game's balancing to make it more fun.

Adeded a health indicator to the game's UI