Added sprite frustum culling
Basically, sprites that aren't in view of the camera are't rendered to the screen. This doesn't do much for simple maps and simple sprites, but for more complex sprites, if shaders are ever added, etc, then the performance gains will be much better.

I adapted much of the code from the original frustum culling file, just slightly modifying it to only focus on the game camera. Seems to work just fine.