Fixed bug where dead players would't die
So my incredible genius showed today when I was messing around with killing players and noticed that sometimes, even though they'd have no hitbox (indicating a health of 0), they simply wouldn't "die" (as in having their sprite disappear and getting a very hilarious death message). I realized that it was because a player would only die of their health was negative, NOT if their health hit exactly 0.