Added the submachine gun
Honestly it's a little hard to balance the SMG, since the high fire rate can obliterate opponents from very close range, but it also has low accuracy and low per bullet damage. I think for now it's in a decent place, but a buff is in order.

Also, I made it a little easier to add both guns and abilities, since now we don't need to mess with uppping the number of items in the randomness code, since the number of items in the enum is counted automatically.