Added the Flamethrower Gun Model
The Flamethrower takes a different approach as a close-corners melee type weapon that heavily relies on head-to-head combat. It has an enormous ammo supply of 200, but it is quickly depleted due to the rapid "shooting" speed that the gun has. This also is a new type of gun to feature multiple material colors, even with some particles being randomly generated.