Added character customization + fixed wasm + slight optimizations
So firstly, players can now set the gun and ability that they want to use on game start. When players set their abilities, they now request to use said ability to the server, however, the server makes the final decision as to what ability the player gets. Basically, when the server receives an ability request, it can send whatever ability back, but it currently chooses to send the ability requested back.

I also fixed WASM builds, since WebGL (and mobile devices) use a slightly different shader language, I had to modify some of the code slightly to compile and run on the web.

I also did some slight optimizations to the codebase as a whole by reducing the mouse position into a Vec2 (which is compatible w. SIMD) and changed the set_ability function to use constants in order to do most of the work at compile time.