Modified flamethrower to shoot multiple flames at the same time.
Instead of sending one strong flame at a time, the flamethrower now sends out many weaker flames. I also gave the flamethrower a slight buff to damage, though it's more difficult to be able to kill two players before reloading. Btw @Susorodni check this for an example on how to make it so a gun shoots multiple bullets at a time.