Made player damage much more consistent w. damage messages
Initially, I was against using damage messages, since it adds more complexity to the game's networking model, but my god it works so much better. Basically, players only take damage if the player who they gave damage to said they did (the honor system of taking damage). Is this prone to abuse, of course. Is it the most fair way that I could think of, yes. Obviously the best solution would be having the server handle all that and deciding based on previous player positions, but I don't feel like doing that. Oh also made the shader loading slightly more concise (it has to be changed in one place instead of 2).