Games can now be won, better player initialization
Firstly, the extremely difficult task of adding like a dozen lines of code so that games are now winnable was done. When a player hits a certain number of kills, the game displays a little text saying the game is over. What should probably happen next is the player is given the choice to head back to the player screen or to continue playing on the server.

Also made it so players initially spawn in w. 0 health and invisible, and that they're only joined in when a player actually goes online.