Added damage indicators
Whenever a player does damage to another player, the amount of damage (rounded) is shown on the screen where the player got hit and fades after 2 seconds. If the player is killed by a bullet, then damage is shown as red.

Due to the way damage currently works in the game, it's possible that on your screen it says you hit the player, but on their screen you didn't. Such is life.