Added dashing
Players now have an ability called "dashing" (binded to the E key by default), where they just move quickly for a fraction of a second, with a 3 second cooldown. I'm not sure if I should make it so that when a player is dashing, they're immune to damage or something, but right now I think that it adds more movement options to the game and just seems fun.