Added instructions for maps + slight improvements to release builds
Firstly, I added information on how maps work in this game, including both the tiny details of how the map format works, and more general information on how to set it up.

Secondly, I added some more flags to release builds of the game to make it just a tiny bit faster.