Added in game settings menu + using Kira for audio
Firstly, I added an in game settings menu that lets you switch guns and abilities in the middle of a match. The downside to this (for the player switching) is that your abilities are reset as are your guns.

I also switched from using the built in audio library that Bevy provides, to (my fork of) bevy_kira, a library that uses Kira to play audio. While we're only using its most basic features now, the fact that it can do much, much more than the built-in Bevy equivalent, as well as working on WASM, means that it's a much better choice imo.