Added proper collision detection + rayon
Fixes #24. Basically, every time an object wants to move a certain distance, it checks that the object wouldn't collide with anything in between (unless the object is moving at over 500 pixels, since then that's a player phasing). The downside to this is that it runs very slowly (as it has to perform over 20 times the calculations).

To compensate for this, I started using Rayon to split the tasks up with multithreading, which significantly increases the speed at which the collision function runs (though unfortuanetly, I can't currently figure out how to run Rayon on WASM, though [this](https://docs.rs/wasm-bindgen-rayon/1.0.1/wasm_bindgen_rayon/) is a good guide. I artshould probably start applying Rayon to the other resource heavy part of the code, move_objects, though since performance is fine now, I'll save that for a future commit.