Massive speed improvements at no cost to iterative compiles
The game now runs at 200 fps (without using V-sync of course) with or without a release build. I wanna say that my optimizations to other tiny stuff in the game was the main reason (though it did increase the frame rate for sure), but 99% of the increase was just increasing the optimization level. Thankfully, however, optimizing all other packages doesn't effect the speed of iterative compiles.