Added descriptions of guns / abilities + fixed multiplayer
I spent some time to try to describe all the guns, abilities, and perks in a way that someone completely new to the game could generally understand it.

I also pretty easily fixed multiplayer builds. So I'm an idiot and I made it so that you had to specify the IP address and port of the server in 2 places instead of just one, so I fixed that by using the lazy_static crate (to make a global variable) and then just specified it at the top of net.rs. I'm eventually gonna have to change it into a Resource instead (since it'll eventually be made that you can connect to any server), but that's a problem for a different day.