Server now chooses map
Slowly making custom maps easier to add. The server now chooses the map, and it chooses it by the CRC32. I also added the map_to_bin function so that, later, I can make it so that if a player doesn't have a map, the server can send it over the network.