Changed player sprites and how they work
So instead of just having the sprites flip back and forth when you move left and right, I made it so that the sprites rotate towards the mouse cursor as well. I also made it so that enemy sprites and player sprites are different, and changed the net code so that the player sends their rotation instead of if they're flipped left or right.

Also found out about x86_64 versions, and just used v3 since that should fit most modern processors (instead of just manually adding a ton of instructions). It also just so happens to include mine and @Susorodni 's processor thats so crazy lol. When I did this, CPU usage noticeably went down. When the actual game releases with an installer, the installer could just check what x86_64 version the computer is using, and pick one of four.