Changed player sprites and how they work
So instead of just having the sprites flip back and forth when you move left and right, I made it so that the sprites rotate towards the mouse cursor as well. I also made it so that enemy sprites and player sprites are different, and changed the net code so that the player sends their rotation instead of if they're flipped left or right.