Unsafe floating point math optimizations
Made some of the most CPU intensive functions (the move_objects function and the collide function) use fast floating point math, that, in exchange for being less accurate, is much faster. On my machine, on release builds, I noticed speed improvements of 13% in terms of CPU usage.

While I could replace much more of the math with fast float math, a major downside to that is the fact that it's inherently much more inaccurate, so I try to use it relatively sparingly.