Fixes #41, added the Ghost Ability, spawn points, bug fixes
- Lets you phase through walls for up to 5 seconds
- If you are stuck in a wall when the ability ends, you die and re-spawn at a random map spawn point
- While phasing, transparency is lowered by 50%, but players can still shoot you.

While working on the ghost ability, we realized that when you died, you would just respawn into a wall, which sucked, so we made it so that you respawn at specific respawn points.

Also fixed a minor bug in net.rs where a player's ability wouldn't be set properly upon joining, leading to a frustrating 30 minute long bug chase.

Co-authored-by: William Batista <bootlegbilly@protonmail.ch>