Made it so players can choose maps
Most of the code changes were in 257ecfc687856016f65ff1f4349ccd128b4dc6d6, but because of a rebase I did I can't edit the commit :(. Basically, maps now have names, and when a player loads a game, the server sets the map. If the player doesn't have the map the server requests, the game just crashes (for now). I plan to use the to_bin function to send the map over the network, but that's for a later date.