Fully added choosing maps, fixed web builds, code cleanup + docs
SO for some reason, Git decided to delete all the changes I made to menus.rs and to setup_systems.rs, but kept the networking code changes needed for choosing maps?? Whatever, easy fix. I tested it w 3 maps, and with swapping maps around in weird ways, and I can safely say that the game does load maps perfectly fine with this change.

Also started using my fork of bevy_webgl2 since the latest Bevy git broke some stuff for that too. Sent a PR fix, hope for the best.

Also added some more documentation to some of the code, and renamed some of it so that it'd be easier to understand exactly what it is. For example, I changed the ShaderPhasing to Alpha, since what that value really does is just set the transparency (alpha) of the player sprites.