Significantly improved speed of object collisions
I thought it couldn't be any faster, but it can...

Rust arrays do bounds checking whenever an item in an array is accessed, and I totally forgot about that. While it usually isn't very much of a drawback, it can be significant when doing many many many bounds checks. Iterators don't have the bounds checking, so they're significantly faster to use.
While it may seem like the game got slower on this commit, it's just because I reduced the optimization level for the game library, so that it would compile faster on dev builds. Trust me when I say it's faster.