Fixed players getting stuck on walls
This seemingly simple fix took like a week to do. Basically, now players won't just stop moving if they can't move in a direction, and instead will continue moving in a direction that they would be typically allowed to. For example, if there is a wall to the left of a player, and the player tries to move down and left, they would just move down (since they can't move left).

One of the current issues with this code is that it increases the number of calcualtions that need to be done, unfortuanetly, though there's probably ways of reducing it.

Also did some clippy lints.