Fixed web builds, slimmed down game lib
Firstly, I specified the binary to the wasm builds, so that they would run correctly.

Secondly, I made it so that the game library doesn't compile the logic.rs file, and making it so instead it's a module of mainrs, since it's entirely unnecessary for the server_cli to compile it.