Switched to lz4_flex for decompressing wasm file
So I've been trying to find ways to make the game in general run faster, and decided that loading the lz4 file and decompressing it had to be part of it. Basically, I wrote some minor bindgen code for lz4_flex over at https://github.com/billyb2/wasm_decompressor/ and then just compiled a release build and copied it over to the project. Now I could add the decompressor code as another rust module, but that would very quickly complicate the build process and only save a few extra kilobytes, so it's not worth the effort.

Oh and also, lz4_flex decompresses much quicker than the original lz4.js file with bas64 wasm, and it also is much smaller. I could maybe add it to the ./tiled/custom.js to replace that old LZ4 code as well, but that's for another day.